Add files via upload
In this project, you will analyze the effectiveness of two different ad campaigns. The company you work for has been experimenting with changing the format of their advertisements. They have been showing the new version (version B) to a subset of users who see their ads on various social media platforms.
You have been tasked with analyzing whether the B version is outperforming the original A version. This is generally known in business analytics as an A/B Test.
You’ll use Python and pandas to merge multiple datasets and perform aggregations to answer questions like:
•	What were the click-through rate performances for each ad version?
•	Did the effectiveness of ads vary across social media platforms?
•	Should we more heavily target mobile, PC, or tablet users in future ad campaigns?